fix: add int for integer

+++ ORM/model.go
@@ -11,7 +11,7 @@ type ORM struct {
 }
 
 type Model struct {
-	ID        int64     `orm-go:"PRIMARY KEY AUTOINCREMENT"`
+	ID        int     `orm-go:"PRIMARY KEY AUTOINCREMENT"`
 	CreatedAt time.Time `orm-go:"DEFAULT CURRENT_TIMESTAMP"`
 }
